getting number of elements in a linked list

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -10,7 +10,7 @@
 
 size_t listint_len(const listint_t *h)
 {
-	size_t p;
+	int p;
 
 	while (h != NULL)
 	{
